Match standard component structure
