+++ packages/commons-ui-core/src/RichTypography/RichTypography.snap.js
